requirements.txt
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
db-dtypes==1.5.0
fastjsonschema==2.21.2
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
nbformat==5.10.4
nbstripout==0.9.1
numpy==2.4.3
packaging==26.0
pandas==2.3.3
platformdirs==4.9.4
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
